dialogue-engine/requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
pydantic>=2.11.0
requests>=2.32.0
PyYAML>=6.0.2
llama-cpp-python>=0.3.2
huggingface-hub>=0.26.0
